requirements.txt
certifi==2025.1.31
google_search_results==2.4.2
openai==1.64.0
pandas==2.2.3
pdfminer==20191125
pdfminer.six==20240706
PyPDF2==3.0.1
python-dotenv==1.0.1
Requests==2.32.3
serpapi==0.1.5
streamlit==1.42.2
